Handle fonts with has_instruction bit set but zero instruction bytes
In one version of the Pacifico font we saw the glyph for é (eacute) had
the has_instructions bit set but a value of zero was read for
instructions length. When this glyph was written back out we only
checked for a non-empty instructions slice to determine whether to write
the optional instruction length value, instead of looking at the flag.
This resulting in a broken font as other tools would attempt to read the
missing instruction length value.

This change makes the glyph writing code check the flags on the glyph to
determine whether to write an instruction length or not.